rename L2TransactionDownloader to LiveL2TransactionDownloader
