+++ packages/backend/src/peripherals/starkware/schema/PerpetualLiveL2TransactionResponse.test.ts
@@ -1,12 +1,12 @@
 import { expect } from 'earl'
 
 import { EXAMPLE_PERPETUAL_TRANSACTIONS } from '../../../test/starkwareData'
-import { PerpetualL2TransactionResponse } from './PerpetualL2TransactionResponse'
+import { PerpetualLiveL2TransactionResponse } from './PerpetualLiveL2TransactionResponse'
 
 describe('PerpetualL2TransactionResponse', () => {
   it('can parse real data', () => {
     const fn = () =>
-      PerpetualL2TransactionResponse.parse(EXAMPLE_PERPETUAL_TRANSACTIONS)
+      PerpetualLiveL2TransactionResponse.parse(EXAMPLE_PERPETUAL_TRANSACTIONS)
     expect(fn).not.toThrow()
   })
 })
